AnalyzAIRR - AIRR-compliant R package for the analysis of bulk Ig/TCR repertoire datasets

AnalyzAIRR is a AIRR-compliant R package developed to analyze bulk Ig/TCR repertoire datasets.

Details

AnalyzAIRR allows a general data exploration to evaluate the homogeneity within defined groups, identify outliers and filter them out. AnalyzAIRR also proposes a set of diversity measures and statistical metrics applicable at any level of granularity. Thus, single-sample repertoire explorations or in-depth cross-comparisons of AIRR datasets can be conducted leading to ready-to-publish visualization graphics. AnalyzAIRR is complemented with a guided workflow to help users in their analytical strategy, and a Shiny web application making it user-friendly for biologists with little or no background in bioinformatics.

See also

Author

Vanessa Mhanna

Gabriel Pires

Grégoire Bohl

Karim El Soufi

Nicolas Tchitchek

David Klatzmann

Adrien Six

Hang-Phuong Pham

Encarnita Mariotti-Ferrandiz