Calculate wilcoxon

calculate wilcoxon tests in all plots

Usage

.safe_kruskal_pairwise(df, group_var, facet_var, color_var, value_var)

Arguments

df

dataframe

group_var

x axis

facet_var

facet group(s)

color_var

color group

Value

dataframe of statistical test results

Details

title Multidimensional Scaling analysis and Principal Component Analysis

description This function can be used to visualize:

- repertoire dissimilarities by performing a multidimensional scaling (MDS) on a pairwise distance matrix calculated between samples on a selected repertoire level, using a specific dissimilarity method.

- differentially expressed repertoire levels calculated using the diffExpGroup function by performing a principal component analysis (PCA).

param x an object of class RepSeqExperiment param level a character specifying the level of the repertoire on which the diversity should be estimated. Should be one of "aaClone","ntClone", "V", "J", "VJ", "ntCDR3" or "aaCDR3". param method a character specifying the distance method to be computed. Should be specified only if the dim_method parameter is set to "MDS", and should be one of the following: "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis." param groupBy a character specifying up to three column names in mData to be used to attribute group colors. param label_colors a list of colors for each factor column in metaData. See plotColors. If NULL, default colors are used. param dim_method a character indicating the dimensional reduction method to be performed. Should be one of "PCA" or "MDS". details Details on the proposed dissimilarity indices can be found in the vegan package:

https://www.rdocumentation.org/packages/vegan/versions/2.4-2/topics/vegdist

export keywords internal examples

data(RepSeqData)

plotDimReduction(x = RepSeqData, level = "V", method = "euclidean", groupBy = "cell_subset", dim_method="MDS")

plotDimReduction(x = RepSeqData, level = "J", groupBy = "cell_subset", dim_method="PCA")

function calculates wilcoxon tests in all plots