Read a single alignment file

parse and filter clonotype tables.

Usage

readAIRR(
  path,
  fileFormat = c("MiXCR", "immunoseq", "MiAIRR"),
  chain = c("TRA", "TRB", "TRG", "TRD", "IGH", "IGK", "IGL"),
  keep.ambiguous = FALSE,
  keep.unproductive = FALSE,
  aa.th = NULL,
  outFiltered = FALSE
)

Arguments

path

a path to an output file of the supported aligners.

fileFormat

a character, the name of the aligner. Should be one of "immunoseq", "MiXCR" or "MiAIRR". Default is MiXCR.

chain

a character, indicates which TCR chain to import: A for the alpha chain and B for the beta chain. Default is A.

keep.ambiguous

a boolean, whether ambiguous sequences (containing a STOP codon) should be kept. Default is FALSE.

keep.unproductive

a boolean, whether unproductive sequences (Euclidean dividion of aa length by 3 > 0) should be kept. Default is FALSE.

aa.th

an interger, indicates the maximum number of amino acids deviating from the mean length to tolerate. Default is NULL.

outFiltered

write in the directory path a data frame containing the columns of the input clonotype table and the filterd out reads as rows. Default path is getwd().

Value

a filtered clonotype table.

Details

this function is a wrapper of parse functions & filterClones

Examples

l <- list.files(system.file(file.path('extdata/MiAIRR'),
                     package = 'AnalyzAIRR'),
                     full.names = TRUE)

dataset <- readAIRR(path = l[1],
                          fileFormat = "MiAIRR",
                          chain = "TRA",
                          keep.ambiguous = FALSE,
                          keep.unproductive = FALSE,
                         outFiltered = FALSE,
                          aa.th = NULL)