Computation of the Renyi index

This function computes the Renyi values at any repertoire level for all the samples in the RepSeqExperiment object.

The alpha values for which the Renyi is to be estimated can be personalized, thus allowing to focus on certain indices such as the Shannon index for alpha=1 or the Simpson index for alpha=2.

Usage

renyiIndex(
  x,
  alpha = c(0, 0.25, 0.5, 1, 2, 4, 8, 16, 32, 64, Inf),
  level = c("aaClone", "ntClone", "V", "J", "VJ", "ntCDR3", "aaCDR3")
)

Arguments

x

an object of class RepSeqExperiment

alpha

a numerical vector specifying the alpha values to compute. If not specified, the following values are estimated: c(0, 0.25, 0.5, 1, 2, 4, 8, 16, 32, 64, Inf).

level

a character specifying the level of the repertoire to be taken into account when calculate VJ usages. Should be one of "aaClone","ntClone", "V", "J", "VJ", "ntCDR3" or "aaCDR3".

Value

a table with Renyi values calculated for all alpha in each sample.

Details

The Renyi index is a generalization of the Shannon index. It represents the distribution of clonal expansions as a function of the parameter alpha. At alpha=0, it equally considers all species including the rare ones, whereas it up-weighs the abundant species with an increasing value of alpha. Alpha =1 is an approximation of the Shannon index; alpha = 2 corresponds to the Simpson index and alpha=Inf corresponds to the Berger-Parker index. The latter expresses the proportional importance of the most abundant species.

Examples

data(RepSeqData)

renyiIndex(RepSeqData, level = "V", alpha = 1)
#> Key: <variable>
#>    variable tripod-30-813 tripod-30-815 tripod-31-846 tripod-31-848
#>      <fctr>         <num>         <num>         <num>         <num>
#> 1:        1          4.04          4.03          4.06          4.02
#>    tripod-35-970 tripod-35-972 tripod-36-1003 tripod-36-1005
#>            <num>         <num>          <num>          <num>
#> 1:          4.04             4           4.07           4.06