The class RepSeqExperiment

An S4 class object enclosing the adaptive immune repertoire data, and which is used in all the analytical metrics proposed by the AnalyzAIRR package. The RepSeqExperiment object is composed of 4 slots, each containing different information.

[] Extract parts of the RepSeqExperiment object

Usage

assay(object)

assay(object, i, j) <- value

oData(object)

oData(object) <- value

mData(object)

mData(object) <- value

History(object)

History(object) <- value

# S4 method for class 'RepSeqExperiment'
assay(object)

# S4 method for class 'RepSeqExperiment'
assay(object, i, j) <- value

# S4 method for class 'RepSeqExperiment'
oData(object)

# S4 method for class 'RepSeqExperiment'
oData(object) <- value

# S4 method for class 'RepSeqExperiment'
mData(object)

# S4 method for class 'RepSeqExperiment'
mData(object) <- value

# S4 method for class 'RepSeqExperiment'
History(object)

# S4 method for class 'RepSeqExperiment'
History(object) <- value

# S4 method for class 'RepSeqExperiment'
show(object)

# S4 method for class 'RepSeqExperiment'
names(x)

# S4 method for class 'RepSeqExperiment,ANY'
names(x) <- value

# S4 method for class 'RepSeqExperiment,ANY,ANY,ANY'
x[i, j, drop]

is.RepSeqExperiment(x)

Arguments

object

a RepSeqExperiment object

i

indice(s) of clone(s) to extract

j

indice(s) of sample(s) to extract

value

either numeric, character or data frame

x

an object

drop

If TRUE the result is coerced to the lowest possible dimension

Value

an object of class [RepSeqExperiment]

Slots

assayData

a data.table binding all clonotype tables and containing the following columns:

- sample_id: sample names

- ntCDR3: nucleotide CDR3 sequence

- aaCDR3: amino acid CDR3 sequence

- V: Variable gene name

- J: Joining gene name

- VJ: V-J gene combination

- aaClone: Full sequence including the V gene, the CDR3 amino acid sequence and the J gene

- ntClone: Full sequence including the the V gene, CDR3 nucleotide sequence and the J gene

- count: the occurrence of the clone sequence

metaData

a data frame containing sample information specified during the building of the RepSeqExperiment object. Each row represents a sample, and columns the possible information or group that can be attributed to the samples such as the cell population, the donor's age, sex, etc... Summary statistics of the assayData slot are also calculated for each sample, as well as the Chao richness estimation.

otherData

a list of data frames containing sequences that were filtered out from the RepSeqExperiment object in case filtering functions were applied.

History

a data frame registering all operations performed on RepSeqExperiment object, such as the filtering functions and the normalization.