Diversity index

calculate a sublist of diversity indices on gene segments.

Usage

divLevel(
  x,
  index = c("shannon", "simpson", "invsimpson"),
  level = c("VJ", "V", "J"),
  norm = TRUE,
  base = exp(1)
)

Arguments

x

an object of class [RepSeqExperiment]

index

character, the diversity index to calculate. Should be one of "chaowor","shannon","invsimpson","simpson" or "gini".

level

character, the level of the repertoire to estimate. Should be one of "V", "J" or "VJ".

norm

boolean, whether to compute the normalized diversity index, which divides the computed value by the log(S), S being the number of observed species in the repetoire. Default is TRUE.

base

a positive or complex number, the base with respect to which logarithms are computed. Default is e=exp(1).

Value

a data.table

Details

function computes the diversity indices for level VJ, V or J only.

Examples

data(RepSeqData)

divLevel(x = RepSeqData, index = "shannon", level = "V", norm = TRUE)
#> Key: <V>
#>             V tripod-30-813 tripod-30-815 tripod-31-846 tripod-31-848
#>        <char>         <num>         <num>         <num>         <num>
#>   1:    TRAV1         0.884         0.947         0.842         0.947
#>   2:   TRAV10         0.879         0.923         0.856         0.917
#>   3:  TRAV10D         0.842         0.908         0.835         0.905
#>   4:  TRAV10N         0.830         0.946         0.861         0.927
#>   5:   TRAV11         0.869         0.950         0.847         0.942
#>  ---                                                                 
#> 100: TRAV9D-2         0.848         0.957         0.865         0.960
#> 101: TRAV9D-3         0.892         0.956         0.844         0.946
#> 102: TRAV9D-4         0.934         0.947         0.869         0.941
#> 103: TRAV9N-3         0.907         0.931         0.862         0.926
#> 104: TRAV9N-4         0.920         0.942         0.892         0.947
#>      tripod-35-970 tripod-35-972 tripod-36-1003 tripod-36-1005
#>              <num>         <num>          <num>          <num>
#>   1:         0.833         0.942          0.945          0.836
#>   2:         0.885         0.923          0.946          0.908
#>   3:         0.828         0.902          0.927          0.892
#>   4:         0.864         0.932          0.921          0.899
#>   5:         0.792         0.933          0.952          0.892
#>  ---                                                          
#> 100:         0.852         0.910          0.916          0.879
#> 101:         0.734         0.938          0.938          0.885
#> 102:         0.897         0.865          0.833          0.975
#> 103:         0.845         0.925          0.890          0.882
#> 104:         0.852         0.930          0.944          0.909

divLevel(x = RepSeqData, index = "simpson", level = "J", norm = FALSE)
#> Key: <J>
#>          J tripod-30-813 tripod-30-815 tripod-31-846 tripod-31-848
#>     <char>         <num>         <num>         <num>         <num>
#>  1: TRAJ11         0.989         0.998         0.995         0.998
#>  2: TRAJ12         0.994         0.998         0.997         0.998
#>  3: TRAJ13         0.984         0.996         0.993         0.996
#>  4: TRAJ15         0.996         0.999         0.991         0.999
#>  5: TRAJ16         0.915         0.997         0.992         0.997
#>  6: TRAJ17         0.995         0.999         0.997         0.999
#>  7: TRAJ18         0.970         0.999         0.978         0.998
#>  8:  TRAJ2         0.991         0.998         0.993         0.998
#>  9: TRAJ21         0.997         0.999         0.997         0.999
#> 10: TRAJ22         0.996         0.998         0.996         0.998
#> 11: TRAJ23         0.995         0.999         0.998         0.999
#> 12: TRAJ24         0.995         0.998         0.994         0.998
#> 13: TRAJ26         0.997         0.999         0.997         0.999
#> 14: TRAJ27         0.997         0.999         0.997         0.999
#> 15: TRAJ28         0.990         0.998         0.992         0.998
#> 16: TRAJ30         0.995         0.999         0.994         0.999
#> 17: TRAJ31         0.995         0.999         0.997         0.999
#> 18: TRAJ32         0.996         0.999         0.998         0.999
#> 19: TRAJ33         0.995         0.999         0.992         0.999
#> 20: TRAJ34         0.997         0.999         0.997         0.999
#> 21: TRAJ35         0.992         0.998         0.993         0.998
#> 22: TRAJ37         0.998         0.999         0.996         0.999
#> 23: TRAJ38         0.992         0.998         0.987         0.998
#> 24: TRAJ39         0.996         0.999         0.994         0.999
#> 25:  TRAJ4         0.994         0.999         0.995         0.999
#> 26: TRAJ40         0.996         0.999         0.995         0.999
#> 27: TRAJ41         0.000         0.444         0.000         0.000
#> 28: TRAJ42         0.995         0.999         0.988         0.999
#> 29: TRAJ43         0.994         0.999         0.994         0.999
#> 30: TRAJ44         0.997         0.998         0.995         0.998
#> 31: TRAJ45         0.988         0.998         0.993         0.999
#> 32: TRAJ47         0.995         0.998         0.994         0.998
#> 33: TRAJ48         0.994         0.999         0.997         0.999
#> 34: TRAJ49         0.996         0.999         0.993         0.999
#> 35:  TRAJ5         0.982         0.997         0.987         0.997
#> 36: TRAJ50         0.996         0.998         0.991         0.998
#> 37: TRAJ52         0.994         0.999         0.995         0.998
#> 38: TRAJ53         0.995         0.998         0.995         0.998
#> 39: TRAJ56         0.996         0.998         0.996         0.999
#> 40: TRAJ57         0.996         0.999         0.994         0.999
#> 41: TRAJ58         0.991         0.998         0.981         0.998
#> 42:  TRAJ6         0.988         0.999         0.995         0.998
#> 43:  TRAJ7         0.994         0.998         0.993         0.998
#> 44:  TRAJ9         0.995         0.999         0.997         0.999
#>          J tripod-30-813 tripod-30-815 tripod-31-846 tripod-31-848
#>     tripod-35-970 tripod-35-972 tripod-36-1003 tripod-36-1005
#>             <num>         <num>          <num>          <num>
#>  1:         0.971         0.998          0.994          0.990
#>  2:         0.996         0.998          0.997          0.994
#>  3:         0.974         0.997          0.995          0.987
#>  4:         0.993         0.999          0.998          0.982
#>  5:         0.978         0.997          0.994          0.915
#>  6:         0.992         0.999          0.997          0.996
#>  7:         0.990         0.998          0.997          0.996
#>  8:         0.982         0.996          0.990          0.975
#>  9:         0.995         0.999          0.998          0.995
#> 10:         0.996         0.998          0.997          0.993
#> 11:         0.989         0.999          0.998          0.997
#> 12:         0.992         0.998          0.994          0.978
#> 13:         0.993         0.999          0.998          0.996
#> 14:         0.995         0.999          0.998          0.993
#> 15:         0.984         0.998          0.993          0.985
#> 16:         0.986         0.999          0.997          0.994
#> 17:         0.985         0.999          0.999          0.995
#> 18:         0.995         0.999          0.996          0.995
#> 19:         0.993         0.999          0.997          0.994
#> 20:         0.979         0.999          0.998          0.997
#> 21:         0.985         0.997          0.988          0.981
#> 22:         0.997         0.999          0.998          0.995
#> 23:         0.989         0.998          0.990          0.988
#> 24:         0.994         0.999          0.997          0.995
#> 25:         0.992         0.999          0.997          0.984
#> 26:         0.996         0.999          0.998          0.996
#> 27:         0.000         0.000          0.000          0.000
#> 28:         0.994         0.999          0.997          0.994
#> 29:         0.981         0.999          0.997          0.995
#> 30:         0.993         0.998          0.995          0.991
#> 31:         0.981         0.998          0.992          0.983
#> 32:         0.974         0.998          0.990          0.983
#> 33:         0.987         0.999          0.994          0.991
#> 34:         0.989         0.999          0.997          0.995
#> 35:         0.967         0.997          0.993          0.981
#> 36:         0.989         0.998          0.994          0.993
#> 37:         0.995         0.998          0.994          0.990
#> 38:         0.979         0.997          0.993          0.993
#> 39:         0.991         0.998          0.995          0.993
#> 40:         0.989         0.999          0.996          0.976
#> 41:         0.994         0.998          0.991          0.983
#> 42:         0.995         0.998          0.996          0.987
#> 43:         0.986         0.998          0.996          0.992
#> 44:         0.996         0.999          0.998          0.995
#>     tripod-35-970 tripod-35-972 tripod-36-1003 tripod-36-1005